Post galeria

+++ app/Http/Controllers/PostController.php
@@ -91,12 +91,15 @@ class PostController extends Controller
    **/
   public function create(Request $request): RedirectResponse
   {
-    dd($request->thumb);
+    if (!$request->has('thumb')) {
+      $request->merge(['thumb' => null]);
+    }
 
-    $request->validate([
+    $request->validate(
+      [
         'titulo' => ['required', 'string', 'max:255'],
         'conteudo' => ['required_if:tipo,noticia', 'string'],
-        'thumb' => ['required_if:tipo,galeria', 'image', 'mimes:jpg,png,jpeg'],
+        'thumb' => ['required',  'sometimes:string', 'sometimes:image', 'sometimes:mimes:jpg,png,jpeg'],
         'data_publicacao' => ['required', 'date'],
       ],
       [
@@ -111,18 +114,41 @@ class PostController extends Controller
       ]
     );
 
-    if(Post::where('slug', Str::slug($request->get('titulo'), '-'))){
+
+
+    if (Post::where('slug', Str::slug($request->get('titulo'), '-'))->exists()) {
       return redirect()->back()->withInput()->with('error', 'Outro post com esse título já existe');
     }
 
     if ($request->hasFile('thumb')) { //thumb
-      foreach($request->thumb as $image) {
-        $originName = $request->file('thumb')->getClientOriginalName();
+      $originName = $request->file('thumb')->getClientOriginalName();
+      $fileName = pathinfo($originName, PATHINFO_FILENAME);
+      $fileName = str_replace(' ', '-', $fileName);
+      $extension = $request->file('thumb')->getClientOriginalExtension();
+      $fileName = $fileName . '_' . time() . '.' . $extension;
+      $request->file('thumb')->move(public_path('post-media'), $fileName);
+
+      // Redimensionar e codificar a imagem para 'jpg' com 75% do tamanho original
+      $img = Image::make(public_path('post-media/' . $fileName));
+      if ($img->height() > 750) {
+        $img->resize(null, 750, function ($constraint) {
+          $constraint->aspectRatio();
+        });
+      }
+      $img->encode('jpg', 75);
+      $img->save(public_path('post-media/' . $fileName));
+
+      $image = $fileName;
+    }
+    // fotos galeria
+    if ($request->hasFile('imagens')) {
+      foreach ($request->file('imagens') as $file) {
+        $originName = $file->getClientOriginalName();
         $fileName = pathinfo($originName, PATHINFO_FILENAME);
         $fileName = str_replace(' ', '-', $fileName);
-        $extension = $request->file('thumb')->getClientOriginalExtension();
+        $extension = $file->getClientOriginalExtension();
         $fileName = $fileName . '_' . time() . '.' . $extension;
-        $request->file('thumb')->move(public_path('post-media'), $fileName);
+        $file->move(public_path('post-media'), $fileName);
 
         // Redimensionar e codificar a imagem para 'jpg' com 75% do tamanho original
         $img = Image::make(public_path('post-media/' . $fileName));
@@ -134,9 +160,17 @@ class PostController extends Controller
         $img->encode('jpg', 75);
         $img->save(public_path('post-media/' . $fileName));
 
-        $image = 'post-media/' . $fileName;
+        $imagePath =  $fileName;
+
+        // Adicione o caminho da imagem na tabela de postmedia
+        $postMedia = PostMedia::create([
+          'slug_post' => Str::slug($request->get('titulo'), '-'),
+          'caminho_media' => $imagePath,
+        ]);
       }
     }
+    // fotos galeria
+
     if (session()->has('tempPastas')) {
       $tempPastas = session()->get('tempPastas');
       $conteudo = $request->get('conteudo');
@@ -163,6 +197,7 @@ class PostController extends Controller
               'slug_post' => Str::slug($request->get('titulo'), '-'),
               'caminho_media' => $file->getFilename(),
             ]);
+
             $postMedia->save();
           }
         }
@@ -200,7 +235,8 @@ class PostController extends Controller
    **/
   public function noticiaInsert(Post $post): View
   {
-    return view('painel.post.insert', ['post' => $post, 'tipo' => 'noticia']);
+    $postMedia = PostMedia::where('slug_post', $post->slug)->get();
+    return view('painel.post.insert', ['post' => $post, 'postMedia' => $postMedia,  'tipo' => 'noticia']);
   }
 
   /**
@@ -211,7 +247,8 @@ class PostController extends Controller
    **/
   public function galeriaInsert(Post $post): View
   {
-    return view('painel.post.insert', ['post' => $post, 'tipo' => 'galeria']);
+    $postMedia = PostMedia::where('slug_post', $post->slug)->get();
+    return view('painel.post.insert', ['post' => $post, 'postMedia' => $postMedia, 'tipo' => 'galeria']);
   }
 
 
@@ -224,11 +261,16 @@ class PostController extends Controller
    **/
   public function update(Request $request, Post $post): RedirectResponse
   {
+    if (!$request->has('thumb')) {
+      $request->merge(['thumb' => null]);
+    }
+
+
     $request->validate(
       [
         'titulo' => ['required', 'string', 'max:255'],
         'conteudo' => ['required_if:tipo,noticia', 'string'],
-        'thumb' => ['required_if:tipo,galeria', 'image', 'mimes:jpg,png,jpeg'],
+        'thumb' => ['required',  'sometimes:string', 'sometimes:image', 'sometimes:mimes:jpg,png,jpeg'],
         'data_publicacao' => ['required', 'date'],
       ],
       [
@@ -243,9 +285,14 @@ class PostController extends Controller
       ]
     );
 
-    if( $request->get('titulo') != $post->titulo && Post::where('slug', Str::slug($request->get('titulo'), '-'))){
-      return redirect()->back()->withInput()->with('error', 'Outro post com esse título já existe');
+    //verifica no update se o titulo foi alterado para um que já exista na base.
+    if ($post->slug != Str::slug($request->get('titulo'), '-')) {
+      if (Post::where('slug', Str::slug($request->get('titulo'), '-'))->exists()) {
+        return redirect()->back()->withInput()->with('error', 'Outro post com esse título já existe');
+      }
     }
+
+
 
     if ($request->hasFile('thumb')) {
       $originName = $request->file('thumb')->getClientOriginalName();
@@ -265,15 +312,49 @@ class PostController extends Controller
       $img->encode('jpg', 75);
       $img->save(public_path('post-media/' . $fileName));
 
-      $image = 'post-media/' . $fileName;
+      $image = $fileName;
 
       // deleta arquivo anterior
-      if (File::exists(public_path($post->thumb))) {
-        File::delete(public_path($post->thumb));
+      if (File::exists(public_path('post-media/' . $post->thumb))) {
+        File::delete(public_path('post-media/' . $post->thumb));
       }
     } else {
       $image = $post->thumb;
     }
+
+    // fotos galeria
+    if ($request->hasFile('imagens')) {
+      foreach ($request->file('imagens') as $file) {
+        $originName = $file->getClientOriginalName();
+        $fileName = pathinfo($originName, PATHINFO_FILENAME);
+        $fileName = str_replace(' ', '-', $fileName);
+        $extension = $file->getClientOriginalExtension();
+        $fileName = $fileName . '_' . time() . '.' . $extension;
+        $file->move(public_path('post-media'), $fileName);
+
+        // Redimensionar e codificar a imagem para 'jpg' com 75% do tamanho original
+        $img = Image::make(public_path('post-media/' . $fileName));
+        if ($img->height() > 750) {
+          $img->resize(null, 750, function ($constraint) {
+            $constraint->aspectRatio();
+          });
+        }
+        $img->encode('jpg', 75);
+        $img->save(public_path('post-media/' . $fileName));
+
+        $imagePath =  $fileName;
+
+
+        // Adicione o caminho da imagem na tabela de postmedia
+        $postMedia = PostMedia::create([
+          'slug_post' => Str::slug($request->get('titulo'), '-'),
+          'caminho_media' => $imagePath,
+
+        ]);
+      }
+    }
+    // fotos galeria
+
     if (session()->has('tempPastas')) {
       $tempPastas = session()->get('tempPastas');
       $conteudo = $request->get('conteudo');
@@ -305,6 +386,7 @@ class PostController extends Controller
       // limpa a session
       session()->forget('tempPastas');
     }
+
     $post->update([
       'titulo' => $request->get('titulo'),
       'slug' => Str::slug($request->get('titulo'), '-'),
@@ -320,19 +402,41 @@ class PostController extends Controller
     //     e em seguida deletar o registro do arquivo na $postMedia
 
     // todos os caminho_media relacionados a este post
-    $postMediaItems = PostMedia::where('slug_post', $post->slug)->get();
-    foreach ($postMediaItems as $postMediaItem) {
-      // procura se tem dentro do conteudo
-      if (strpos($post->conteudo, $postMediaItem->caminho_media) === false) {
-        // deleta o arquivo do diretório
-        $fileToDelete = public_path('post-media/' . $postMediaItem->caminho_media);
-        if (File::exists($fileToDelete)) {
-          File::delete($fileToDelete);
+    if ($post->tipo == 'noticia') {
+      $postMediaItems = PostMedia::where('slug_post', $post->slug)->get();
+      foreach ($postMediaItems as $postMediaItem) {
+        // procura se tem dentro do conteudo
+        if (strpos($post->conteudo, $postMediaItem->caminho_media) === false) {
+          // deleta o arquivo do diretório
+          $fileToDelete = public_path('post-media/' . $postMediaItem->caminho_media);
+          if (File::exists($fileToDelete)) {
+            File::delete($fileToDelete);
+          }
+          // deleta o registro da tabela post_media
+          $postMediaItem->delete();
         }
-        // deleta o registro da tabela post_media
-        $postMediaItem->delete();
       }
     }
+    //deleta fotos que foram descartados no update da galeria e adicionados a listanegra
+    if ($request->has('deleteList')) {
+      $deleteList = explode(',', $request->input('deleteList'));
+      foreach ($deleteList as $id) {
+        // Recupera o registro PostMedia pelo id
+        $postMediaItem = PostMedia::find($id);
+
+        if ($postMediaItem) {
+          // Deleta o arquivo do diretório
+          $fileToDelete = public_path('post-media/' . $postMediaItem->caminho_media);
+          if (File::exists($fileToDelete)) {
+            File::delete($fileToDelete);
+          }
+
+          // Deleta o registro da tabela post_media
+          $postMediaItem->delete();
+        }
+      }
+    }
+
     if (!$post) {
       return redirect()->back()->with('error', 'Ocorreu um erro!');
     }
@@ -350,14 +454,14 @@ class PostController extends Controller
    **/
   public function thumbDelete(Post $post): RedirectResponse
   {
-      // deleta arquivo anterior
-      if (File::exists(public_path($post->thumb))) {
-          File::delete(public_path($post->thumb));
-      }
+    // deleta arquivo anterior
+    if (File::exists(public_path($post->thumb))) {
+      File::delete(public_path($post->thumb));
+    }
 
-      $post->update(['thumb' => null]);
+    $post->update(['thumb' => null]);
 
-      return redirect()->back()->with('update-success', 'thumb removido');
+    return redirect()->back()->with('update-success', 'thumb removido');
   }
 
   /**
@@ -370,8 +474,8 @@ class PostController extends Controller
   public function delete(Request $request, Post $post, PostMedia $postMedia): RedirectResponse
   {
     //deleta a thumb
-    if (File::exists(public_path($post->thumb))) {
-      File::delete(public_path($post->thumb));
+    if (File::exists(public_path('post-media/' . $post->thumb))) {
+      File::delete(public_path('post-media/' . $post->thumb));
     }
     // todos os caminho_media relacionados a este post
     $postMediaItems = PostMedia::where('slug_post', $post->slug)->get();
@@ -387,7 +491,12 @@ class PostController extends Controller
     }
     $post->delete();
 
-    return redirect()->route('noticia-index')->with('update-success', 'Post removido');;
+
+
+    if ($post->tipo == 'noticia') {
+      return redirect()->route('noticia-index')->with('update-success', 'Noticia adicionada');
+    }
+    return redirect()->route('galeria-index')->with('update-success', 'Galeria adicionada');
   }
 
   /**
@@ -444,11 +553,13 @@ class PostController extends Controller
   /**
    * mostra pagina da slug da noticia/galeria no site
    *
+   
    * @return View
    **/
   public function show($slug)
   {
+    $postMedia = PostMedia::where('slug_post', $slug)->get();
     $post = Post::where('slug', $slug)->firstOrFail();
-    return view('site.pages.slug-da-noticia', ['post' => $post]);
+    return view('site.pages.slug-da-noticia', ['post' => $post, 'postMedia' => $postMedia]);
   }
 }
